release: patch version bump (4.1.3 -> 4.1.4)
Further fixes to the 'browser' specification.  Drops the single-line
"simple" spec in favour of the object version.  In this case we give a
hint to downstream bundlers to replace instances of "urbit-ob" with the
browser bundle we distribute.

(This helped fix an issue that occurred while running urbit-ob as a
dependency of urbit-key-generation in Bridge.)